JdbcUtils.getCatalystType maps TINYINT to IntegerType instead of ByteType

+++ sql/core/src/test/scala/org/apache/spark/sql/jdbc/JDBCSuite.scala
@@ -572,7 +572,7 @@ class JDBCSuite extends QueryTest
     assert(rows.length === 1)
     assert(rows(0).getInt(0) === 1)
     assert(rows(0).getBoolean(1) === false)
-    assert(rows(0).getInt(2) === 3)
+    assert(rows(0).getByte(2) === 3)
     assert(rows(0).getInt(3) === 4)
     assert(rows(0).getLong(4) === 1234567890123L)
   }
